+++ yangvoodoo/VoodooNode.py
@@ -1,4 +1,4 @@
-from yangvoodoo import Cache, Common, Errors, TemplateNinja, Types
+from yangvoodoo import Cache, Common, Errors, Types
 
 
 class Context:
@@ -270,38 +270,7 @@ class Empty():
 
 class ContainingNode(Node):
 
-    def _from_template(self, template, **kwargs):
-        """
-        TODO: decide if this is good here or if it should move to session.
-
-        This is misleading because the template still applies from root (although
-        potentially we cahnge templater to handle an abritary node)
-
-        """
-        templater = TemplateNinja.TemplateNinja()
-        templater.from_template(self, template, **kwargs)
-
-    def _from_xmlstr(self, xmlstr):
-        """
-        TODO: decide if this is good here or if it should move to session.
-
-        This is misleading because the template still applies from root (although
-        potentially we cahnge templater to handle an abritary node)
-
-        """
-        templater = TemplateNinja.TemplateNinja()
-        templater.from_xmlstr(self, xmlstr)
-
-    def _to_xmlstr(self):
-        """
-        TODO: decide if this is good here or if it should move to session.
-
-        This is misleading because the template still applies from root (although
-        potentially we cahnge templater to handle an abritary node)
-        """
-
-        templater = TemplateNinja.TemplateNinja()
-        return templater.to_xmlstr(self._context.dal.dump_xpaths())
+    pass
 
 
 class Choice(Node):
@@ -686,68 +655,6 @@ class Root(ContainingNode):
         context = self.__dict__['_context']
         return "VoodooTopNode{} YANG Module: " + context.module
 
-    def from_template(self, template, **kwargs):
-        """
-        Process a template with a number of data nodes. The result of processing the template
-        with Jinja2 must be a valid XML document.
-
-        Example:
-            session.from_template(root, 'template1.xml')
-
-            Process the template from the the path specified (i.e. template1.xml)
-            In the Jinja2 templates the root object is available as 'root.'
-
-            session.from_template(root, 'template1.xml', data_a=root.morecomplex)
-
-            In this case the Jinja2 template will receive both 'data_a' as the subset of
-            data at /morecomplex and root.
-
-        The path to the template may be specified as a relative path (to where the python
-        process is running (i.e. os.getcwd) or an exact path.
-
-        IMPORTANT to note is that variable and logic is processed in the template based upon
-        the data available at the time, then the result of the entire template is applied.
-        To be clear consdiering this template
-
-        This template isn't quite the same as NETCONF - we don't have any namespaces in it,
-        instead the contents are wrapped in a node matching the name of the yang module.
-            <integrationtest>
-                <simpleleaf>HELLO</simpleleaf>
-                <default>{{ root.simpleleaf }}</default>
-            </integrationtest>
-
-            root.simpleleaf = 'GOODBYE'
-            session.from_template(root, 'hello-goodbye.xml')
-
-        The resulting value for simpleleaf will be 'GOODBYE' not hello.
-
-        """
-
-        templater = TemplateNinja.TemplateNinja()
-        templater.from_template(self, template, **kwargs)
-
-    def to_xmlstr(self):
-        """
-        Return an XML Template from this node.
-        """
-
-        templater = TemplateNinja.TemplateNinja()
-        return templater.to_xmlstr(self._context.dal.dump_xpaths())
-
-    def from_xmlstr(self, xmlstr):
-        """
-        This template isn't quite the same as NETCONF - we don't have any namespaces in it,
-        instead the contents are wrapped in a node matching the name of the yang module.
-            xmlstr = "<integrationtest><simpleleaf>HELLO</simpleleaf></integrationtest>"
-
-            root.simpleleaf = 'GOODBYE'
-            session.from_xmlstr(root, xmlstr)
-
-        """
-
-        templater = TemplateNinja.TemplateNinja()
-        templater.from_xmlstr(self, xmlstr)
-
 
 class SuperRoot:
 
@@ -777,40 +684,3 @@ class SuperRoot:
 
     def __repr__(self):
         return "VoodooSuperRoot{}"
-
-    def from_template(self, template, **kwargs):
-        """
-        Process a template with a number of data nodes. The result of processing the template
-        with Jinja2 must be a valid XML document.
-
-        Example:
-            session.from_template(root, 'template1.xml')
-
-            Process the template from the the path specified (i.e. template1.xml)
-            In the Jinja2 templates the root object is available as 'root.'
-
-            session.from_template(root, 'template1.xml', data_a=root.morecomplex)
-
-            In this case the Jinja2 template will receive both 'data_a' as the subset of
-            data at /morecomplex and root.
-
-        The path to the template may be specified as a relative path (to where the python
-        process is running (i.e. os.getcwd) or an exact path.
-
-        IMPORTANT to note is that variable and logic is processed in the template based upon
-        the data available at the time, then the result of the entire template is applied.
-        To be clear consdiering this template
-            <integrationtest>
-                <simpleleaf>HELLO</simpleleaf>
-                <default>{{ root.simpleleaf }}</default>
-            </integrationtest>
-
-            root.simpleleaf = 'GOODBYE'
-            session.from_template(root, 'hello-goodbye.xml')
-
-        The resulting value for simpleleaf will be 'GOODBYE' not hello.
-
-        """
-
-        templater = TemplateNinja.TemplateNinja()
-        templater.from_template(self._node, template, **kwargs)