+++ CHANGELOG.md
@@ -99,3 +99,4 @@
    - Allow data to be loaded as libyang TRUSTED (i.e. don't enfoce when/must/mandatory checks). (default for merges)
    - Implement validate for stubly using lyd_validation from libyang.
    - If a DAL returns None for a number looking type return 0 instead.
+   - Deprecated TemplateNinja